Remove per-mob values of sprint speed, sneak speed, etc.
This is something I'd be interested in revisiting at some point, but not worthwhile to pursue right now. The basic reason it didn't work is because when the mob was being created, the makeBrain method was getting called BEFORE the FelinePredator had been initialized. So the values only got correctly set for mobs that already existed when makeBrain was called, so if you restarted the world, any existing mobs would have the correct values.